refactor(rules): split heading_levels into impl/test and align with TS
